requirements.txt
asgiref==3.6.0
Django==3.2.16
django-appconf==1.0.5
django-bootstrap5==22.2
django-imagekit==4.1.0
django-shortcuts==1.6
pilkit==2.0
Pillow==9.4.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
